+++ CHANGELOG.md
@@ -1,1 +1,6 @@
-## 
+## 0.1.0
+
+- **Added**
+  - Minify `Simple Requests` module to minify the chat area it takes up
+  - Colorize `PF2e HUD`'s `Persistent HUD` to have color
+  - Add User Color to Message Header